Remaining MCO2 fixes

+++ polygon-digital-carbon/src/utils/DailyKlimaRetireSnapshot.ts
@@ -8,8 +8,8 @@ export function loadOrCreateDailyKlimaRetireSnapshot(id: string): DailyKlimaReti
     retire.timestamp = BigInt.zero()
     retire.pool = new Bytes(0x0)
     retire.credit = new Bytes(0x0)
-    retire.amount = BigInt.fromString('0')
-    retire.feeAmount = BigInt.fromString('0')
+    retire.amount = BigDecimal.fromString('0')
+    retire.feeAmount = BigDecimal.fromString('0')
   }
 
   return retire as DailyKlimaRetireSnapshot
